backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pinata-sdk
web3
requests